✨Feat: avgStar + user Management

+++ src/main/java/com/mega/_NY/review/controller/ReviewController.java
@@ -31,8 +31,8 @@ public class ReviewController {
     //리뷰 생성
     @PostMapping
     public ResponseEntity<ReviewDTO> createReview(@RequestBody ReviewDTO reviewDTO){
-         ReviewDTO createdReview = reviewService.createReview(reviewDTO);
-         return ResponseEntity.ok(reviewDTO);
+        ReviewDTO createdReview = reviewService.createReview(reviewDTO);
+        return ResponseEntity.ok(reviewDTO);
     }
 
     //특정 id 리뷰 조회
@@ -43,12 +43,18 @@ public class ReviewController {
     }
 
     //사용자 모든 리뷰 조회
-    @GetMapping("/user/{userId}")
-    public ResponseEntity<Page<ReviewDTO>> getReviewsByUserId(@PathVariable Long userId, Pageable pageable) {
+    @GetMapping("/user")
+    public ResponseEntity<Page<ReviewDTO>> getReviewsByUserId(
+            @RequestParam(value="page", defaultValue = "0") int page,
+            @RequestParam(value="size", defaultValue="10") int size
+    ) {
         if (!isUser()) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
 
+        Long userId = userService.getLoginUser().getId();
+
+        Pageable pageable = PageRequest.of(page, size);
         Page<ReviewDTO> reviews = reviewService.getReviewsByUserId(userId, pageable);
         return ResponseEntity.ok(reviews);
     }
@@ -59,7 +65,7 @@ public class ReviewController {
             @PathVariable Long itemId,
             @RequestParam(value="page", defaultValue = "0") int page,
             @RequestParam(value="size", defaultValue="10") int size
-            ) {
+    ) {
 
         Pageable pageable = PageRequest.of(page, size);
         Page<ReviewDTO> reviews = reviewService.getReviewsByItemId(itemId, pageable);
@@ -69,7 +75,18 @@ public class ReviewController {
     //리뷰 수정
     @PutMapping("/{reviewId}")
     public ResponseEntity<ReviewDTO> updateReview(@PathVariable Long reviewId,
-                                                   @RequestBody ReviewDTO reviewDTO){
+                                                  @RequestBody ReviewDTO reviewDTO){
+        if (!isUser()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        ReviewDTO existReview = reviewService.getReview(reviewId);
+
+        // 리뷰 작성자와 사용자가 일치하는지 확인
+        if (!existReview.getUserId().equals(userService.getLoginUser().getId())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         ReviewDTO updateReviewDTO = reviewService.updateReview(reviewId, reviewDTO);
         return ResponseEntity.ok(updateReviewDTO);
     }
@@ -77,6 +94,18 @@ public class ReviewController {
     //리뷰 삭제
     @DeleteMapping("/{reviewId}")
     public ResponseEntity<ReviewDTO> deleteReview(@PathVariable Long reviewId){
+
+        if (!isUser()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        ReviewDTO existReview = reviewService.getReview(reviewId);
+
+        // 리뷰 작성자와 사용자가 일치하는지 확인
+        if (!existReview.getUserId().equals(userService.getLoginUser().getId())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         reviewService.deleteReview(reviewId);
         return ResponseEntity.noContent().build();
